agents/claude-hello-world/pyproject.toml
[project]
name = "claude-hello-world"
version = "1.0.0"
description = "A simple agent demonstrating Claude Agent SDK capabilities"
requires-python = ">=3.12"
dependencies = [
    "claude-agent-sdk",
    "shared",
]

[tool.uv.sources]
shared = { path = "../../shared" }

[project.optional-dependencies]
dev = [
    "python-dotenv",
]

[project.scripts]
run-agent = "run_agent:main"
